refactor: split tests out of dict_word_metadata
